--- test_scripts/Polices/user_consent_of_Policies/ATF_Data_consent_device_not_in_LPT.lua
@@ -1,94 +1,0 @@
----------------------------------------------------------------------------------------------
--- Requirement summary: 
---    [Policies]: user data concent - device that does not have records in "user_consent_records"
---
--- Description: 
---     User activates the application which runs on the device that does not have records in "user_consent_records"-> "device" sub-section of Local PT
---     1. Used preconditions:
--- 			Connect device
--- 			Register application
---
---     2. Performed steps
---		    Activate application 
---
--- Expected result:
---     PoliciesManager must initiate getting User`s data consent (that is, User`s permission for using the mobile device`s connection for Policy Table exchange)
-----------------------------------------------------------------------------------------------------
---[[ General Settings for configuration ]]
-Test = require('user_modules/connecttest_resumption')
-require('cardinalities')
-local mobile_session = require('mobile_session')
-
---[[ General configuration parameters ]]
-config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
-
---[[ Required Shared libraries ]]
-local commonFunctions = require ('user_modules/shared_testcases/commonFunctions')
-local commonSteps = require('user_modules/shared_testcases/commonSteps')
-local commonTestCases = require('user_modules/shared_testcases/commonTestCases')
-require('user_modules/AppTypes')
-
---[[ Preconditions ]]
-commonSteps:DeleteLogsFileAndPolicyTable()
-
-function Test:Precondition_ConnectDevice()
-	commonTestCases:DelayedExp(2000)
-	self:connectMobile()
-	EXPECT_HMICALL("BasicCommunication.UpdateDeviceList",
-					{
-						deviceList = {
-							{
-								id = config.deviceMAC,
-								isSDLAllowed = false,
-								name = "127.0.0.1",
-								transportType = "WIFI"
-							}
-						}
-					}
-	):Do(function(_,data)
-		self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
-	end)
-	:Times(AtLeast(1))
-end
-
-function Test:Precondition_Register_app()
-	commonTestCases:DelayedExp(3000)
-	self.mobileSession = mobile_session.MobileSession(self, self.mobileConnection)
-	self.mobileSession:StartService(7)
-	:Do(function()
-		local correlationId = self.mobileSession:SendRPC("RegisterAppInterface", config.application1.registerAppInterfaceParams)
-		EXPECT_HMINOTIFICATION("BasicCommunication.OnAppRegistered")
-		:Do(function(_,data)
-			self.HMIAppID = data.params.application.appID
-		end)
-		self.mobileSession:ExpectResponse(correlationId, { success = true, resultCode = "SUCCESS" })
-		self.mobileSession:ExpectNotification("OnHMIStatus", {hmiLevel = "NONE", audioStreamingState = "NOT_AUDIBLE", systemContext = "MAIN"})
-	end)
-end
-
---[[ Test ]]
-commonFunctions:newTestCasesGroup("Test")
-
-function Test:Check_user_data_consent_is_requested()
-	local RequestId = self.hmiConnection:SendRequest("SDL.ActivateApp", { appID = self.applications["Test Application"]})
-	EXPECT_HMIRESPONSE(RequestId, {result = { code = 0, device = { id = config.deviceMAC, name = "127.0.0.1" }, isSDLAllowed = false, isPermissionsConsentNeeded = true, method ="SDL.ActivateApp", priority ="NONE"}})
-	:Do(function(_,data)
-		if data.result.isSDLAllowed ~= false then
-			commonFunctions:userPrint(31, "Error: wrong behavior of SDL - device not from LPT, should be consented on HMI")
-		else
-			local RequestIdGetMes = self.hmiConnection:SendRequest("SDL.GetUserFriendlyMessage", {language = "EN-US", messageCodes = {"DataConsent"}})
-			EXPECT_HMIRESPONSE(RequestIdGetMes)
-			:Do(function()
-				self.hmiConnection:SendNotification("SDL.OnAllowSDLFunctionality", 
-				{allowed = true, source = "GUI", device = {id = config.deviceMAC, name = "127.0.0.1"}})	
-			end)
-		end
-	end)
-end
-
---[[ Postconditions ]]
-commonFunctions:newTestCasesGroup("Postconditions")
-
-function Test.Postcondition_SDLForceStop()
-  commonFunctions:SDLForceStop()
-end
